+++ test/api.test.js
@@ -1,6 +1,7 @@
 const fhirpath = require('../src/fhirpath');
 const r4_model = require('../fhir-context/r4');
 const _ = require('lodash');
+const {FP_DateTime, FP_Quantity} = require('../src/types');
 
 describe('compile', () => {
   it('should accept a model object', () => {
@@ -36,6 +37,30 @@ describe('compile', () => {
     );
     let result = execExpression({}, {partOfResource});
     expect(result).toStrictEqual([true]);
+  });
+
+  it('should resolve values which have internal data types to strings by default', () => {
+    let f = fhirpath.compile('@2018-02-18T12:23:45-05:00');
+    expect(f({})).toStrictEqual(['2018-02-18T12:23:45-05:00']);
+
+    f = fhirpath.compile("2.0 'cm'");
+    expect(f({})).toStrictEqual(["2 'cm'"]);
+  });
+
+  it('should not resolve values which have internal data types to strings when options.resolveInternalTypes is false', () => {
+    let f = fhirpath.compile(
+      '@2018-02-18T12:23:45-05:00',
+      null,
+      {resolveInternalTypes: false}
+    );
+    expect(f({})).toStrictEqual([new FP_DateTime('2018-02-18T12:23:45-05:00')]);
+
+    f = fhirpath.compile(
+      "2.0 'cm'",
+      null,
+      {resolveInternalTypes: false}
+    );
+    expect(f({})).toStrictEqual([new FP_Quantity(2, "'cm'")]);
   });
 });
 
@@ -84,5 +109,50 @@ describe('evaluate', () => {
     );
     expect(result).toEqual(['1', '2', '3', '4']);
     expect(someVar).toStrictEqual(someVarOrig);
-  })
+  });
+
+  it('should resolve values which have internal data types to strings by default', () => {
+    expect(
+      fhirpath.evaluate({}, '@2018-02-18T12:23:45-05:00')
+    ).toStrictEqual(['2018-02-18T12:23:45-05:00']);
+
+    expect(
+      fhirpath.evaluate({}, "2.0 'cm'")
+    ).toStrictEqual(["2 'cm'"]);
+  });
+
+  it('should not resolve values which have internal data types to strings when options.resolveInternalTypes is false', () => {
+    expect(
+      fhirpath.evaluate(
+        {},
+        '@2018-02-18T12:23:45-05:00',
+        null,
+        null,
+        { resolveInternalTypes: false })
+    ).toStrictEqual([new FP_DateTime('2018-02-18T12:23:45-05:00')]);
+
+    expect(
+      fhirpath.evaluate(
+        {},
+        "2.0 'cm'",
+        null,
+        null,
+        { resolveInternalTypes: false }
+      )
+    ).toStrictEqual([new FP_Quantity(2, "'cm'")]);
+  });
+});
+
+describe('resolveInternalTypes', () => {
+  it('should resolve values which have internal data types to strings', () => {
+    expect(
+      fhirpath.resolveInternalTypes([
+        new FP_DateTime('2020-02-18T12:23:45-05:00'),
+        new FP_Quantity(1, "'cm'")
+      ])
+    ).toStrictEqual([
+      '2020-02-18T12:23:45-05:00',
+      "1 'cm'"
+    ]);
+  });
 });